Add DXL broker example and adjusted otherFields schema
The otherFields schema is now basically a subset of the 'payload'
schema (which uses Swagger's 'schema' directly).

This commit also creates a definition for stringArray under the OpenDXL
API schema which allows for the array to be empty. draft-04 of the core
JSON schema required at least one element to be in the array.
stringArray in later versions of the core JSON schema allow for the
array to be empty so it should be possible to remove this in the future.