Change service metadata to have generic instead of instance data
Previously, service metadata in an OpenDXL API document was expressed
with values for the metadata items, making the data instance-specific.
This commit changes the metadata definition to provide a description and
name for the metadata items but not values for the items. (The values
would be used in an actual service instance registration with the DXL
fabric.)